Bug #52137: remove ability to select basesystem in USS ApplianceWizard.js
Theerefor the radiobutton to select basesystem on system role page
is removed. The Waring page for the basesystem is not needed anymore.
Also the _isRoleBaseSystem function and its uses are
removed, because this function now always would return false.
All checks for 'basesystem' and places there this is set are removed.