Bug #49981: Test passed
Not an extended attribute, but still it prooves, that the mapping file
is working. The mapping file contains a mapping from the ucs-attribute
'city' to the ad attribute 'organisation'. While the test is running one
can watch under windows how the user has the attribute and the assertion
in the end of the test validates that as well. Test passed.