Bug #52273: don't count down session timeout second wise
restore behavior before git:5cdc3849dd5e70055a2c35fdd950532aaac8cb43.

The UMC-Web-Server created a timer which counts down the session second
wise for each session to check if it can be destroyed.

When 1000 uses are logged in currently, every seconds 1000
timer-callbacks are called.

Instead we should register a timer at the time when the session should
be destroyed.
This was not possible prior to python-notifier running in Python 2
because time.time() instead of time.monotonic() was used. Changing the
system time (e.g. via rdate in the system setup) then caused sessions being
closed while they are still valid.